refactor(utilities): extract common utilities into separate files
Move FirstMouseHostingView, paste functionality, hotkey utils, and notifications into dedicated utility files
Centralize notification name definitions in Notifications.swift
Deprecate old notification name in UpdateService
Remove duplicate code from OverlayWindowController and AppFocusService
