+++ Utilities/FirstMouseHostingView.swift
@@ -1,0 +1,6 @@
+import SwiftUI
+import AppKit
+
+final class FirstMouseHostingView<Content: View>: NSHostingView<Content> {
+    override func acceptsFirstMouse(for event: NSEvent?) -> Bool { true }
+}
