+++ Views/SettingsView.swift
@@ -120,23 +120,7 @@ struct SettingsView: View {
     }
     
     private var hotKeyDescription: String {
-        var components: [String] = []
-        
-        if settings.hotkeyModifiers.containsCmd { components.append("⌘") }
-        if settings.hotkeyModifiers.containsCtrl { components.append("⌃") }
-        if settings.hotkeyModifiers.containsAlt { components.append("⌥") }
-        if settings.hotkeyModifiers.containsShift { components.append("⇧") }
-        
-        // Best-effort letter from keyCode (only supports A-Z here)
-        let letterMap: [UInt32: String] = [
-            0x00:"A",0x0B:"B",0x08:"C",0x02:"D",0x0E:"E",0x03:"F",0x05:"G",0x04:"H",0x22:"I",0x26:"J",
-            0x28:"K",0x25:"L",0x2E:"M",0x2D:"N",0x1F:"O",0x23:"P",0x0C:"Q",0x0F:"R",0x01:"S",0x11:"T",
-            0x20:"U",0x09:"V",0x0D:"W",0x07:"X",0x10:"Y",0x06:"Z"
-        ]
-        let key = letterMap[settings.hotkeyKeyCode] ?? "V"
-        components.append(key)
-        
-        return components.joined()
+        HotkeyUtils.description(keyCode: settings.hotkeyKeyCode, modifiers: settings.hotkeyModifiers)
     }
     
     private func startCapturingHotkey() {
